refactor: delete bridge/session_logs.py and models/reflections.py shims
Both modules were pure re-export shims left over from earlier refactors.

- bridge/session_logs.py re-exported from agent/session_logs.py and had zero
  production importers; its only consumer was a test asserting the re-export.
- models/reflections.py re-exported ReflectionIgnore / PRReviewAudit. Its three
  production importers now import from models.reflection_ignore and
  models.pr_review_audit directly.

Also hardens three mock patch targets in test_reflections_package.py. The
production imports are function-local, so patching the shim path worked only by
accident of import order. Repointing production without repointing these would
have turned the patches into silent no-ops that still passed while hitting real
Redis. Each now asserts the mock was actually called, so a stale patch target
fails loudly instead of silently degrading into an integration test.

Closes #2872
