--- bridge/session_logs.py
@@ -1,13 +1,0 @@
-"""
-Per-session log snapshots at key lifecycle transitions.
-
-BACKWARD COMPATIBILITY: This module re-exports from agent.session_logs,
-which is the canonical location. Import from agent.session_logs for new code.
-"""
-
-# Re-export everything from the canonical location
-from agent.session_logs import (  # noqa: F401
-    SESSION_LOGS_DIR,
-    cleanup_old_snapshots,
-    save_session_snapshot,
-)
