+++ docs/features/reflections.md
@@ -840,8 +840,8 @@ worker-role install gate, cutover ordering, and the `/dashboard.json`
 |---------|-------------|
 | `tail -f logs/worker.log` | Stream worker logs (includes reflection scheduler output) |
 | `curl -s localhost:8500/dashboard.json` | Full system state including reflection status |
-| `python -c "from models.reflections import ReflectionIgnore; [print(f'{e.pattern}') for e in ReflectionIgnore.get_active()]"` | View active ignore entries |
-| `python -c "from models.reflections import ReflectionIgnore; ReflectionIgnore.add_ignore('pattern', reason='why', days=14)"` | Add an ignore entry |
+| `python -c "from models.reflection_ignore import ReflectionIgnore; [print(f'{e.pattern}') for e in ReflectionIgnore.get_active()]"` | View active ignore entries |
+| `python -c "from models.reflection_ignore import ReflectionIgnore; ReflectionIgnore.add_ignore('pattern', reason='why', days=14)"` | Add an ignore entry |
 
 ## Key Files
 
@@ -864,7 +864,6 @@ worker-role install gate, cutover ordering, and the `/dashboard.json`
 | `models/reflection.py` | Reflection state model (per-reflection Redis tracking) |
 | `models/reflection_ignore.py` | ReflectionIgnore: auto-fix suppression with TTL-based expiry |
 | `models/pr_review_audit.py` | PRReviewAudit: PR review finding deduplication |
-| `models/reflections.py` | Re-export shim: `ReflectionIgnore`, `PRReviewAudit` |
 | `scripts/reflections_report.py` | GitHub issue creation module (was used by retired `daily_report`) |
 | `scripts/update/env_sync.py` | `sync_reflections_yaml()`: writes the real-file vault copy on update |
 | `~/Desktop/Valor/projects.json` | Multi-repo project registry |
