+++ docs/features/worker-service.md
@@ -186,7 +186,7 @@ The session execution engine (`agent/agent_session_queue.py`) has zero module-le
 | What | Canonical Location | Re-export Location |
 |------|-------------------|-------------------|
 | `REACTION_SUCCESS/COMPLETE/ERROR` | `agent/constants.py` | `bridge/response.py` |
-| `save_session_snapshot()` | `agent/session_logs.py` | `bridge/session_logs.py` |
+| `save_session_snapshot()` | `agent/session_logs.py` | none (import canonically) |
 
 All 7 lazy `bridge/` imports inside functions are guarded with `try/except` for graceful degradation when running without Telegram.
