--- models/reflections.py
@@ -1,15 +1,0 @@
-"""Backward-compatible re-export shim for reflection models.
-
-Models are in individual files:
-- ReflectionIgnore -> models/reflection_ignore.py
-- PRReviewAudit -> models/pr_review_audit.py
-
-ReflectionRun has been removed (issue #748). The docs_auditor.py
-singleton pattern was migrated to a plain Redis key ('docs_auditor:last_audit_date').
-All other ReflectionRun usage was in the deleted scripts/reflections.py monolith.
-"""
-
-from models.pr_review_audit import PRReviewAudit
-from models.reflection_ignore import ReflectionIgnore
-
-__all__ = ["ReflectionIgnore", "PRReviewAudit"]
